Move show selectors to the form to track required
[#152464351]
