+++ app/javascript/containers/screenings/AllegationsShowContainer.jsx
@@ -2,9 +2,11 @@ import {connect} from 'react-redux'
 import AllegationShow from 'views/allegations/AllegationShow'
 import {
   getFormattedAllegationsSelector,
+} from 'selectors/screening/allegationShowSelectors'
+import {
   getAllegationsRequiredValueSelector,
   getAllegationsAlertErrorMessageSelector,
-} from 'selectors/screening/allegationShowSelectors'
+} from 'selectors/screening/allegationsFormSelectors'
 import {getScreeningIsReadOnlySelector} from 'selectors/screeningSelectors'
 
 const mapStateToProps = (state, {toggleMode}) => {
