+++ app/javascript/selectors/screening/allegationsFormSelectors.js
@@ -2,10 +2,15 @@ import {createSelector} from 'reselect'
 import {Map, List, fromJS} from 'immutable'
 import {getScreeningSelector} from 'selectors/screeningSelectors'
 import nameFormatter from 'utils/nameFormatter'
-
+import {siblingAtRiskHasRequiredComplementaryAllegations} from 'utils/allegationsHelper'
+import ALLEGATION_TYPES from 'enums/AllegationTypes'
 const FLATTEN_LEVEL = 1
+
 const getPeopleSelector = (state) => state.get('participants', List())
 const getAllegationsFormSelector = (state) => state.get('allegationsForm', List())
+export const getAllegationTypesSelector = () => (
+  fromJS(ALLEGATION_TYPES.map((type) => ({label: type.value, value: type.value})))
+)
 
 const getVictimsSelector = createSelector(
   getPeopleSelector,
@@ -17,11 +22,16 @@ const getPerpetratorsSelector = createSelector(
   (people) => people.filter((person) => person.get('roles', List()).includes('Perpetrator'))
 )
 
-const getAllegationsToSaveSelector = createSelector(
+const getAllegationsWithTypesSelector = createSelector(
   getAllegationsFormSelector,
   (allegations) => allegations.filterNot((allegation) => (
     allegation.get('allegationTypes').filterNot((type) => type === '').isEmpty()
-  )).map((allegation) => Map({
+  ))
+)
+
+const getAllegationsToSaveSelector = createSelector(
+  getAllegationsWithTypesSelector,
+  (allegations) => allegations.map((allegation) => Map({
     victim_id: allegation.get('victimId'),
     perpetrator_id: allegation.get('perpetratorId'),
     allegation_types: allegation.get('allegationTypes'),
@@ -56,4 +66,25 @@ export const getFormattedAllegationsSelector = createSelector(
       })
     )).flatten(FLATTEN_LEVEL)
   )
+)
+
+export const getAllegationsRequiredValueSelector = createSelector(
+  getScreeningSelector,
+  (screening) => screening.get('screening_decision') === 'promote_to_referral'
+)
+
+export const getAllegationsAlertErrorMessageSelector = createSelector(
+  getScreeningSelector,
+  getAllegationsRequiredValueSelector,
+  getAllegationsFormSelector,
+  getAllegationsWithTypesSelector,
+  (screening, required, allegations, allegationsWithTypes) => {
+    if (!siblingAtRiskHasRequiredComplementaryAllegations(allegations)) {
+      return 'Any allegations of Sibling at Risk must be accompanied by another allegation.'
+    } else if (required && allegationsWithTypes.isEmpty()) {
+      return 'Any report that is promoted for referral must include at least one allegation.'
+    } else {
+      return undefined
+    }
+  }
 )
